Use the correct type size

+++ src/c/sk_font.cpp
@@ -141,7 +141,7 @@ float sk_font_measure_text(const sk_font_t* font, const void* text, size_t byteL
     return AsFont(font)->measureText(text, byteLength, (SkTextEncoding)encoding, AsRect(bounds), AsPaint(paint));
 }
 
-size_t sk_font_break_text(const sk_font_t* font, const void* text, int byteLength, sk_text_encoding_t encoding, float maxWidth, float* measuredWidth, const sk_paint_t* paint) {
+size_t sk_font_break_text(const sk_font_t* font, const void* text, size_t byteLength, sk_text_encoding_t encoding, float maxWidth, float* measuredWidth, const sk_paint_t* paint) {
     return AsFont(font)->breakText(text, byteLength, (SkTextEncoding)encoding, maxWidth, measuredWidth, AsPaint(paint));
 }
